chore: combine stm32-utils and stm32-common
