+++ apps/battery-node/src/main.c
@@ -2,11 +2,14 @@
 #include <string.h>
 
 #include "app.h"
+#include "clock.h"
+#include "error.h"
+#include "flash.h"
 #include "mayor.h"
 #include "peripherals.h"
 #include "ports.h"
 #include "postmaster-hal.h"
-#include "utils.h"
+#include "print.h"
 
 // FreeRTOS
 #include "FreeRTOS.h"
@@ -14,8 +17,6 @@
 
 // Hardware
 static peripherals_t *peripherals;
-
-void system_clock_init(void);
 
 // FreeRTOS
 #define BATTERY_MONITOR_TASK_PERIOD 200
@@ -79,8 +80,8 @@ int main(void) {
   // Reset of all peripherals, Initializes the Flash interface and the Systick.
   HAL_Init();
 
-  if (FlashRWInit() != APP_OK) {
-    Error_Handler();
+  if (flash_init() != APP_OK) {
+    error();
   }
 
   system_clock_init();
@@ -100,54 +101,13 @@ int main(void) {
 
   task_init();
 
-  Print(&peripherals->huart1, "Starting application...\r\n");
+  print(&peripherals->huart1, "Starting application...\r\n");
 
   // Start scheduler
   vTaskStartScheduler();
 
   // We should never get here as control is now taken by the scheduler.
   while (1) {
-  }
-}
-
-void system_clock_init(void) {
-  RCC_OscInitTypeDef RCC_OscInitStruct = {0};
-  RCC_ClkInitTypeDef RCC_ClkInitStruct = {0};
-  RCC_PeriphCLKInitTypeDef PeriphClkInit = {0};
-
-  /** Initializes the RCC Oscillators according to the specified parameters
-   * in the RCC_OscInitTypeDef structure.
-   */
-  RCC_OscInitStruct.OscillatorType = RCC_OSCILLATORTYPE_HSE;
-  RCC_OscInitStruct.HSEState = RCC_HSE_ON;
-  RCC_OscInitStruct.HSIState = RCC_HSI_ON;
-  RCC_OscInitStruct.PLL.PLLState = RCC_PLL_ON;
-  RCC_OscInitStruct.PLL.PLLSource = RCC_PLLSOURCE_HSE;
-  RCC_OscInitStruct.PLL.PLLMUL = RCC_PLL_MUL6;
-  RCC_OscInitStruct.PLL.PREDIV = RCC_PREDIV_DIV1;
-  if (HAL_RCC_OscConfig(&RCC_OscInitStruct) != HAL_OK) {
-    Error_Handler();
-  }
-
-  /** Initializes the CPU, AHB and APB buses clocks
-   */
-  RCC_ClkInitStruct.ClockType = RCC_CLOCKTYPE_HCLK | RCC_CLOCKTYPE_SYSCLK |
-                                RCC_CLOCKTYPE_PCLK1 | RCC_CLOCKTYPE_PCLK2;
-  RCC_ClkInitStruct.SYSCLKSource = RCC_SYSCLKSOURCE_PLLCLK;
-  RCC_ClkInitStruct.AHBCLKDivider = RCC_SYSCLK_DIV1;
-  RCC_ClkInitStruct.APB1CLKDivider = RCC_HCLK_DIV2;
-  RCC_ClkInitStruct.APB2CLKDivider = RCC_HCLK_DIV1;
-
-  if (HAL_RCC_ClockConfig(&RCC_ClkInitStruct, FLASH_LATENCY_2) != HAL_OK) {
-    Error_Handler();
-  }
-  PeriphClkInit.PeriphClockSelection =
-      RCC_PERIPHCLK_USART1 | RCC_PERIPHCLK_I2C1 | RCC_PERIPHCLK_ADC12;
-  PeriphClkInit.Usart1ClockSelection = RCC_USART1CLKSOURCE_PCLK2;
-  PeriphClkInit.Adc12ClockSelection = RCC_ADC12PLLCLK_DIV1;
-  PeriphClkInit.I2c1ClockSelection = RCC_I2C1CLKSOURCE_SYSCLK;
-  if (HAL_RCCEx_PeriphCLKConfig(&PeriphClkInit) != HAL_OK) {
-    Error_Handler();
   }
 }
 
@@ -253,7 +213,7 @@ void mayor_init(void) {
   };
 
   if (ck_mayor_init(&mayor) != CK_OK) {
-    Error_Handler();
+    error();
   }
 }
 
@@ -269,7 +229,9 @@ void task_init(void) {
 
 void HAL_ADC_ConvCpltCallback(ADC_HandleTypeDef *hadc) {
   (void)hadc;
-  NotifyTask(power_measure_task);
+  BaseType_t xHigherPriorityTaskWoken = pdFALSE;
+  vTaskNotifyGiveFromISR(power_measure_task, &xHigherPriorityTaskWoken);
+  portYIELD_FROM_ISR(xHigherPriorityTaskWoken);
 }
 
 void HAL_GPIO_EXTI_Callback(uint16_t GPIO_Pin) {
@@ -281,12 +243,16 @@ void HAL_GPIO_EXTI_Callback(uint16_t GPIO_Pin) {
 
 void battery_monitor_timer(TimerHandle_t xTimer) {
   (void)xTimer;
-  NotifyTask(battery_monitor_task);
+  BaseType_t xHigherPriorityTaskWoken = pdFALSE;
+  vTaskNotifyGiveFromISR(battery_monitor_task, &xHigherPriorityTaskWoken);
+  portYIELD_FROM_ISR(xHigherPriorityTaskWoken);
 }
 
 void power_measure_timer(TimerHandle_t xTimer) {
   (void)xTimer;
-  NotifyTask(power_measure_task);
+  BaseType_t xHigherPriorityTaskWoken = pdFALSE;
+  vTaskNotifyGiveFromISR(power_measure_task, &xHigherPriorityTaskWoken);
+  portYIELD_FROM_ISR(xHigherPriorityTaskWoken);
 }
 
 void power_measure(void *argument) {
@@ -329,13 +295,13 @@ void power_measure(void *argument) {
 
     // Send the documents
     if (ck_send_document(cell_folder->folder_no) != CK_OK) {
-      Error_Handler();
+      error();
     }
     if (ck_send_document(reg_out_current_folder->folder_no) != CK_OK) {
-      Error_Handler();
+      error();
     }
     if (ck_send_document(vbat_out_current_folder->folder_no) != CK_OK) {
-      Error_Handler();
+      error();
     }
   }
 }
