+++ apps/battery-node/src/peripherals.c
@@ -1,8 +1,8 @@
 #include "peripherals.h"
 
+#include "error.h"
 #include "ports.h"
 #include "stm32f3xx_hal.h"
-#include "utils.h"
 
 #define DMA1_Channel1_IRQ_PRIORITY 5
 #define DMA2_Channel1_IRQ_PRIORITY 5
@@ -43,14 +43,14 @@ void adc1_init(void) {
   hadc1->Init.LowPowerAutoWait = DISABLE;
   hadc1->Init.Overrun = ADC_OVR_DATA_OVERWRITTEN;
   if (HAL_ADC_Init(hadc1) != HAL_OK) {
-    Error_Handler();
+    error();
   }
 
   /** Configure the ADC multi-mode
    */
   multimode.Mode = ADC_MODE_INDEPENDENT;
   if (HAL_ADCEx_MultiModeConfigChannel(hadc1, &multimode) != HAL_OK) {
-    Error_Handler();
+    error();
   }
 
   /** Configure Regular Channel
@@ -62,7 +62,7 @@ void adc1_init(void) {
   sConfig.OffsetNumber = ADC_OFFSET_NONE;
   sConfig.Offset = 0;
   if (HAL_ADC_ConfigChannel(hadc1, &sConfig) != HAL_OK) {
-    Error_Handler();
+    error();
   }
 
   /** Configure Regular Channel
@@ -70,7 +70,7 @@ void adc1_init(void) {
   sConfig.Channel = ADC_CHANNEL_2;
   sConfig.Rank = ADC_REGULAR_RANK_2;
   if (HAL_ADC_ConfigChannel(hadc1, &sConfig) != HAL_OK) {
-    Error_Handler();
+    error();
   }
 
   /** Configure Regular Channel
@@ -78,7 +78,7 @@ void adc1_init(void) {
   sConfig.Channel = ADC_CHANNEL_3;
   sConfig.Rank = ADC_REGULAR_RANK_3;
   if (HAL_ADC_ConfigChannel(hadc1, &sConfig) != HAL_OK) {
-    Error_Handler();
+    error();
   }
 
   /** Configure Regular Channel
@@ -86,11 +86,11 @@ void adc1_init(void) {
   sConfig.Channel = ADC_CHANNEL_4;
   sConfig.Rank = ADC_REGULAR_RANK_4;
   if (HAL_ADC_ConfigChannel(hadc1, &sConfig) != HAL_OK) {
-    Error_Handler();
+    error();
   }
 
   if (HAL_ADCEx_Calibration_Start(hadc1, ADC_SINGLE_ENDED) != HAL_OK) {
-    Error_Handler();
+    error();
   }
 }
 
@@ -120,7 +120,7 @@ void adc2_init(void) {
   hadc2->Init.LowPowerAutoWait = DISABLE;
   hadc2->Init.Overrun = ADC_OVR_DATA_OVERWRITTEN;
   if (HAL_ADC_Init(hadc2) != HAL_OK) {
-    Error_Handler();
+    error();
   }
 
   /** Configure Regular Channel
@@ -132,7 +132,7 @@ void adc2_init(void) {
   sConfig.OffsetNumber = ADC_OFFSET_NONE;
   sConfig.Offset = 0;
   if (HAL_ADC_ConfigChannel(hadc2, &sConfig) != HAL_OK) {
-    Error_Handler();
+    error();
   }
 
   /** Configure Regular Channel
@@ -140,7 +140,7 @@ void adc2_init(void) {
   sConfig.Channel = ADC_CHANNEL_2;
   sConfig.Rank = ADC_REGULAR_RANK_2;
   if (HAL_ADC_ConfigChannel(hadc2, &sConfig) != HAL_OK) {
-    Error_Handler();
+    error();
   }
 
   /** Configure Regular Channel
@@ -148,7 +148,7 @@ void adc2_init(void) {
   sConfig.Channel = ADC_CHANNEL_3;
   sConfig.Rank = ADC_REGULAR_RANK_3;
   if (HAL_ADC_ConfigChannel(hadc2, &sConfig) != HAL_OK) {
-    Error_Handler();
+    error();
   }
 
   /** Configure Regular Channel
@@ -156,11 +156,11 @@ void adc2_init(void) {
   sConfig.Channel = ADC_CHANNEL_11;
   sConfig.Rank = ADC_REGULAR_RANK_4;
   if (HAL_ADC_ConfigChannel(hadc2, &sConfig) != HAL_OK) {
-    Error_Handler();
+    error();
   }
 
   if (HAL_ADCEx_Calibration_Start(hadc2, ADC_SINGLE_ENDED) != HAL_OK) {
-    Error_Handler();
+    error();
   }
 }
 
@@ -184,7 +184,7 @@ void can_init(void) {
   hcan->Init.ReceiveFifoLocked = DISABLE;
   hcan->Init.TransmitFifoPriority = DISABLE;
   if (HAL_CAN_Init(hcan) != HAL_OK) {
-    Error_Handler();
+    error();
   }
 }
 
@@ -205,19 +205,19 @@ void i2c1_init(void) {
   hi2c1->Init.GeneralCallMode = I2C_GENERALCALL_DISABLE;
   hi2c1->Init.NoStretchMode = I2C_NOSTRETCH_DISABLE;
   if (HAL_I2C_Init(hi2c1) != HAL_OK) {
-    Error_Handler();
+    error();
   }
 
   /** Configure Analogue filter
    */
   if (HAL_I2CEx_ConfigAnalogFilter(hi2c1, I2C_ANALOGFILTER_ENABLE) != HAL_OK) {
-    Error_Handler();
+    error();
   }
 
   /** Configure Digital filter
    */
   if (HAL_I2CEx_ConfigDigitalFilter(hi2c1, 0) != HAL_OK) {
-    Error_Handler();
+    error();
   }
 }
 
@@ -244,7 +244,7 @@ void spi1_init(void) {
   hspi1->Init.CRCLength = SPI_CRC_LENGTH_DATASIZE;
   hspi1->Init.NSSPMode = SPI_NSS_PULSE_ENABLE;
   if (HAL_SPI_Init(hspi1) != HAL_OK) {
-    Error_Handler();
+    error();
   }
 }
 
@@ -266,7 +266,7 @@ void uart1_init(void) {
   huart1->Init.OneBitSampling = UART_ONE_BIT_SAMPLE_DISABLE;
   huart1->AdvancedInit.AdvFeatureInit = UART_ADVFEATURE_NO_INIT;
   if (HAL_UART_Init(huart1) != HAL_OK) {
-    Error_Handler();
+    error();
   }
 }
 
@@ -373,7 +373,7 @@ void adc1_dma_init(ADC_HandleTypeDef* hadc) {
   hdma_adc1->Init.Mode = DMA_NORMAL;
   hdma_adc1->Init.Priority = DMA_PRIORITY_LOW;
   if (HAL_DMA_Init(hdma_adc1) != HAL_OK) {
-    Error_Handler();
+    error();
   }
   __HAL_LINKDMA(hadc, DMA_Handle, *hdma_adc1);
 }
@@ -390,7 +390,7 @@ void adc2_dma_init(ADC_HandleTypeDef* hadc) {
   hdma_adc2->Init.Mode = DMA_NORMAL;
   hdma_adc2->Init.Priority = DMA_PRIORITY_LOW;
   if (HAL_DMA_Init(hdma_adc2) != HAL_OK) {
-    Error_Handler();
+    error();
   }
 
   __HAL_LINKDMA(hadc, DMA_Handle, *hdma_adc2);
