+++ apps/io-node/src/main.c
@@ -2,8 +2,11 @@
 #include <string.h>
 
 #include "app.h"
+#include "clock.h"
+#include "error.h"
+#include "flash.h"
 #include "peripherals.h"
-#include "utils.h"
+#include "print.h"
 
 // FreeRTOS
 #include "FreeRTOS.h"
@@ -11,8 +14,6 @@
 
 // Hardware
 static peripherals_t *peripherals;
-
-void system_clock_config(void);
 
 // FreeRTOS
 #define IO_READ_TASK_PERIOD_MS 20
@@ -33,12 +34,12 @@ int main(void) {
   // Reset of all peripherals, Initializes the Flash interface and the Systick.
   HAL_Init();
 
-  if (FlashRWInit() != APP_OK) {
-    Error_Handler();
+  if (flash_init() != APP_OK) {
+    error();
   }
 
   // Configure the system clock
-  system_clock_config();
+  system_clock_init();
 
   // Initialize all configured peripherals
   peripherals = get_peripherals();
@@ -53,58 +54,13 @@ int main(void) {
 
   task_init();
 
-  Print(&peripherals->huart1, "Starting application...\n");
+  print(&peripherals->huart1, "Starting application...\n");
 
   // Start scheduler
   vTaskStartScheduler();
 
   // We should never get here as control is now taken by the scheduler
   while (1) {
-  }
-}
-
-/**
- * @brief System Clock Configuration
- * @retval None
- */
-void system_clock_config(void) {
-  RCC_OscInitTypeDef RCC_OscInitStruct = {0};
-  RCC_ClkInitTypeDef RCC_ClkInitStruct = {0};
-  RCC_PeriphCLKInitTypeDef PeriphClkInit = {0};
-
-  /** Initializes the RCC Oscillators according to the specified parameters
-   * in the RCC_OscInitTypeDef structure.
-   */
-  RCC_OscInitStruct.OscillatorType = RCC_OSCILLATORTYPE_HSE;
-  RCC_OscInitStruct.HSEState = RCC_HSE_ON;
-  RCC_OscInitStruct.HSIState = RCC_HSI_ON;
-  RCC_OscInitStruct.PLL.PLLState = RCC_PLL_ON;
-  RCC_OscInitStruct.PLL.PLLSource = RCC_PLLSOURCE_HSE;
-  RCC_OscInitStruct.PLL.PLLMUL = RCC_PLL_MUL6;
-  RCC_OscInitStruct.PLL.PREDIV = RCC_PREDIV_DIV1;
-  if (HAL_RCC_OscConfig(&RCC_OscInitStruct) != HAL_OK) {
-    Error_Handler();
-  }
-
-  /** Initializes the CPU, AHB and APB buses clocks
-   */
-  RCC_ClkInitStruct.ClockType = RCC_CLOCKTYPE_HCLK | RCC_CLOCKTYPE_SYSCLK |
-                                RCC_CLOCKTYPE_PCLK1 | RCC_CLOCKTYPE_PCLK2;
-  RCC_ClkInitStruct.SYSCLKSource = RCC_SYSCLKSOURCE_PLLCLK;
-  RCC_ClkInitStruct.AHBCLKDivider = RCC_SYSCLK_DIV1;
-  RCC_ClkInitStruct.APB1CLKDivider = RCC_HCLK_DIV2;
-  RCC_ClkInitStruct.APB2CLKDivider = RCC_HCLK_DIV1;
-
-  if (HAL_RCC_ClockConfig(&RCC_ClkInitStruct, FLASH_LATENCY_2) != HAL_OK) {
-    Error_Handler();
-  }
-  PeriphClkInit.PeriphClockSelection =
-      RCC_PERIPHCLK_USART1 | RCC_PERIPHCLK_I2C1 | RCC_PERIPHCLK_ADC12;
-  PeriphClkInit.Usart1ClockSelection = RCC_USART1CLKSOURCE_PCLK2;
-  PeriphClkInit.Adc12ClockSelection = RCC_ADC12PLLCLK_DIV1;
-  PeriphClkInit.I2c1ClockSelection = RCC_I2C1CLKSOURCE_SYSCLK;
-  if (HAL_RCCEx_PeriphCLKConfig(&PeriphClkInit) != HAL_OK) {
-    Error_Handler();
   }
 }
 
@@ -116,12 +72,16 @@ void task_init(void) {
 
 void HAL_ADC_ConvCpltCallback(ADC_HandleTypeDef *hadc) {
   UNUSED(hadc);
-  NotifyTask(io_read_task);
+  BaseType_t xHigherPriorityTaskWoken = pdFALSE;
+  vTaskNotifyGiveFromISR(io_read_task, &xHigherPriorityTaskWoken);
+  portYIELD_FROM_ISR(xHigherPriorityTaskWoken);
 }
 
 void io_read_timer(TimerHandle_t xTimer) {
   UNUSED(xTimer);
-  NotifyTask(io_read_task);
+  BaseType_t xHigherPriorityTaskWoken = pdFALSE;
+  vTaskNotifyGiveFromISR(io_read_task, &xHigherPriorityTaskWoken);
+  portYIELD_FROM_ISR(xHigherPriorityTaskWoken);
 }
 
 void io_read(void *argument) {
