+++ apps/servo-node/src/main.c
@@ -2,8 +2,11 @@
 #include <string.h>
 
 #include "app.h"
+#include "clock.h"
+#include "error.h"
+#include "flash.h"
 #include "peripherals.h"
-#include "utils.h"
+#include "print.h"
 
 // FreeRTOS
 #include "FreeRTOS.h"
@@ -13,8 +16,6 @@
 
 // Hardware
 static peripherals_t *peripherals;
-
-void system_clock_config(void);
 
 // FreeRTOS
 #define PWM_TASK_PERIOD_MS 20
@@ -50,12 +51,12 @@ int main(void) {
   // Reset of all peripherals, Initializes the Flash interface and the Systick.
   HAL_Init();
 
-  if (FlashRWInit() != APP_OK) {
-    Error_Handler();
+  if (flash_init() != APP_OK) {
+    error();
   }
 
   // Configure the system clock
-  system_clock_config();
+  system_clock_init();
 
   // Initialize all configured peripherals
   peripherals = get_peripherals();
@@ -77,61 +78,13 @@ int main(void) {
 
   can_message_queue = xQueueCreate(CAN_MESSAGE_QUEUE_LENGTH, sizeof(CANFrame));
 
-  Print(&peripherals->huart1, "Starting application...\r\n");
+  print(&peripherals->huart1, "Starting application...\r\n");
 
   // Start scheduler
   vTaskStartScheduler();
 
   // We should never get here as control is now taken by the scheduler
   while (1) {
-  }
-}
-
-/**
- * @brief System Clock Configuration
- * @retval None
- */
-void system_clock_config(void) {
-  RCC_OscInitTypeDef RCC_OscInitStruct = {0};
-  RCC_ClkInitTypeDef RCC_ClkInitStruct = {0};
-  RCC_PeriphCLKInitTypeDef PeriphClkInit = {0};
-
-  /** Initializes the RCC Oscillators according to the specified parameters
-   * in the RCC_OscInitTypeDef structure.
-   */
-  RCC_OscInitStruct.OscillatorType = RCC_OSCILLATORTYPE_HSE;
-  RCC_OscInitStruct.HSEState = RCC_HSE_ON;
-  RCC_OscInitStruct.HSIState = RCC_HSI_ON;
-  RCC_OscInitStruct.PLL.PLLState = RCC_PLL_ON;
-  RCC_OscInitStruct.PLL.PLLSource = RCC_PLLSOURCE_HSE;
-  RCC_OscInitStruct.PLL.PLLMUL = RCC_PLL_MUL6;
-  RCC_OscInitStruct.PLL.PREDIV = RCC_PREDIV_DIV1;
-  if (HAL_RCC_OscConfig(&RCC_OscInitStruct) != HAL_OK) {
-    Error_Handler();
-  }
-
-  /** Initializes the CPU, AHB and APB buses clocks
-   */
-  RCC_ClkInitStruct.ClockType = RCC_CLOCKTYPE_HCLK | RCC_CLOCKTYPE_SYSCLK |
-                                RCC_CLOCKTYPE_PCLK1 | RCC_CLOCKTYPE_PCLK2;
-  RCC_ClkInitStruct.SYSCLKSource = RCC_SYSCLKSOURCE_PLLCLK;
-  RCC_ClkInitStruct.AHBCLKDivider = RCC_SYSCLK_DIV1;
-  RCC_ClkInitStruct.APB1CLKDivider = RCC_HCLK_DIV2;
-  RCC_ClkInitStruct.APB2CLKDivider = RCC_HCLK_DIV1;
-
-  if (HAL_RCC_ClockConfig(&RCC_ClkInitStruct, FLASH_LATENCY_2) != HAL_OK) {
-    Error_Handler();
-  }
-  PeriphClkInit.PeriphClockSelection = RCC_PERIPHCLK_USART1 |
-                                       RCC_PERIPHCLK_I2C1 | RCC_PERIPHCLK_I2C3 |
-                                       RCC_PERIPHCLK_TIM1 | RCC_PERIPHCLK_ADC12;
-  PeriphClkInit.Usart1ClockSelection = RCC_USART1CLKSOURCE_PCLK2;
-  PeriphClkInit.Adc12ClockSelection = RCC_ADC12PLLCLK_DIV1;
-  PeriphClkInit.I2c1ClockSelection = RCC_I2C1CLKSOURCE_SYSCLK;
-  PeriphClkInit.I2c3ClockSelection = RCC_I2C3CLKSOURCE_SYSCLK;
-  PeriphClkInit.Tim1ClockSelection = RCC_TIM1CLK_HCLK;
-  if (HAL_RCCEx_PeriphCLKConfig(&PeriphClkInit) != HAL_OK) {
-    Error_Handler();
   }
 }
 
@@ -150,36 +103,44 @@ void task_init(void) {
 
 void HAL_ADC_ConvCpltCallback(ADC_HandleTypeDef *hadc) {
   UNUSED(hadc);
-  NotifyTask(power_measure_task);
+  BaseType_t xHigherPriorityTaskWoken = pdFALSE;
+  vTaskNotifyGiveFromISR(power_measure_task, &xHigherPriorityTaskWoken);
+  portYIELD_FROM_ISR(xHigherPriorityTaskWoken);
 }
 
-void defaultTaskTimer(TimerHandle_t xTimer) {
+void pwm_timer(TimerHandle_t xTimer) {
   UNUSED(xTimer);
-  NotifyTask(pwm_task);
+  BaseType_t xHigherPriorityTaskWoken = pdFALSE;
+  vTaskNotifyGiveFromISR(pwm_task, &xHigherPriorityTaskWoken);
+  portYIELD_FROM_ISR(xHigherPriorityTaskWoken);
 }
 
-void measureTaskTimer(TimerHandle_t xTimer) {
+void power_measure_timer(TimerHandle_t xTimer) {
   UNUSED(xTimer);
-  NotifyTask(power_measure_task);
+  BaseType_t xHigherPriorityTaskWoken = pdFALSE;
+  vTaskNotifyGiveFromISR(power_measure_task, &xHigherPriorityTaskWoken);
+  portYIELD_FROM_ISR(xHigherPriorityTaskWoken);
 }
 
-void mailboxFreeCallback(CAN_HandleTypeDef *_hcan) {
+void mailbox_free_callback(CAN_HandleTypeDef *_hcan) {
   // Only notify when going from 0 free mailboxes to 1 free
   if (HAL_CAN_GetTxMailboxesFreeLevel(_hcan) <= 1) {
-    NotifyTask(can_tx_task);
+    BaseType_t xHigherPriorityTaskWoken = pdFALSE;
+    vTaskNotifyGiveFromISR(can_tx_task, &xHigherPriorityTaskWoken);
+    portYIELD_FROM_ISR(xHigherPriorityTaskWoken);
   }
 }
 
 void HAL_CAN_TxMailbox0CompleteCallback(CAN_HandleTypeDef *_hcan) {
-  mailboxFreeCallback(_hcan);
+  mailbox_free_callback(_hcan);
 }
 
 void HAL_CAN_TxMailbox1CompleteCallback(CAN_HandleTypeDef *_hcan) {
-  mailboxFreeCallback(_hcan);
+  mailbox_free_callback(_hcan);
 }
 
 void HAL_CAN_TxMailbox2CompleteCallback(CAN_HandleTypeDef *_hcan) {
-  mailboxFreeCallback(_hcan);
+  mailbox_free_callback(_hcan);
 }
 
 void can_tx(void *argument) {
@@ -216,7 +177,7 @@ void power_measure(void *argument) {
       "measureTaskTimer", pdMS_TO_TICKS(POWER_MEASURE_TASK_PERIOD_MS),
       pdTRUE,  // Auto reload timer
       NULL,    // Timer ID, unused
-      measureTaskTimer);
+      power_measure_timer);
 
   xTimerStart(xTimer, portMAX_DELAY);
 
@@ -263,7 +224,7 @@ void pwm(void *argument) {
       xTimerCreate("defaultTaskTimer", pdMS_TO_TICKS(PWM_TASK_PERIOD_MS),
                    pdTRUE,  // Auto reload timer
                    NULL,    // Timer ID, unused
-                   defaultTaskTimer);
+                   pwm_timer);
 
   xTimerStart(xTimer, portMAX_DELAY);
   HAL_TIM_PWM_Start(&peripherals->htim1, TIM_CHANNEL_4);
