+++ libs/meson.build
@@ -1,7 +1,6 @@
 subdir('can-kingdom')
 subdir('stm32-common')
 subdir('stm32-postmaster')
-subdir('stm32-utils')
 subdir('third-party')
 
 # Convenience source file list
@@ -9,7 +8,6 @@ common_libs_src = [
   can_kingdom_src,
   stm32_common_src,
   stm32_postmaster_src,
-  stm32_utils_src,
   drivers_src,
   freertos_src,
 ]
@@ -19,7 +17,6 @@ common_libs_inc = [
   can_kingdom_inc,
   stm32_common_inc,
   stm32_postmaster_inc,
-  stm32_utils_inc,
   drivers_inc,
   freertos_inc,
 ]
@@ -28,5 +25,4 @@ common_libs_tidy_files = [
   can_kingdom_tidy_files,
   stm32_common_tidy_files,
   stm32_postmaster_tidy_files,
-  stm32_utils_tidy_files,
 ]
