+++ libs/stm32-common/include/clock.h
@@ -1,0 +1,14 @@
+#ifndef CLOCK_H
+#define CLOCK_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+void system_clock_init(void);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* CLOCK_H */
