+++ libs/stm32-common/meson.build
@@ -2,8 +2,12 @@ stm32_common_inc = include_directories('include')
 
 stm32_common_src = files(
   'src' / 'assert.c',
+  'src' / 'clock.c',
   'src' / 'common-interrupts.c',
+  'src' / 'error.c',
+  'src' / 'flash.c',
   'src' / 'freertos.c',
+  'src' / 'print.c',
   'src' / 'startup_stm32f302xe.s',
   'src' / 'system_stm32f3xx.c',
 )
@@ -16,6 +20,10 @@ stm32_linker_script = configure_file(
 
 stm32_common_tidy_files = files(
   'src' / 'assert.c',
+  'src' / 'clock.c',
   'src' / 'common-interrupts.c',
+  'src' / 'error.c',
+  'src' / 'flash.c',
   'src' / 'freertos.c',
+  'src' / 'print.c',
 )
