+++ libs/stm32-common/src/error.c
@@ -1,0 +1,9 @@
+#include "error.h"
+
+#include "stm32f3xx_hal.h"
+
+void error(void) {
+  __disable_irq();
+  while (1) {
+  }
+}
