--- libs/stm32-utils/meson.build
@@ -1,3 +1,0 @@
-stm32_utils_inc = include_directories('include')
-stm32_utils_src = files('src' / 'utils.c')
-stm32_utils_tidy_files = stm32_utils_src
